fixed a 64/32 mismatch, with thanks to Larsivi

+++ tango/text/stream/StreamIterator.d
@@ -64,7 +64,7 @@ class StreamIterator(T) : InputFilter
 
         ***********************************************************************/
 
-        abstract protected uint scan (void[] data);
+        abstract protected size_t scan (void[] data);
 
         /***********************************************************************
